+++ ntex-rt/src/builder.rs
@@ -309,6 +309,7 @@ impl SystemRunner {
             })
             .await;
 
+        crate::arbiter::run_shutdown_callbacks();
         unsafe {
             crate::remove_all_items();
         }